[hardware] :bug: Introduce instruction counters in the main sequencer
Previous situation:
The main sequencer gives a valid instruction to the lane sequencer.
But this valid instruction remains valid until ALL the Ara units are ready again.
This should not be a problem in principle, as each new instruction that is validated
by the main sequencer goes to one unit only, and if this instruction makes the unit non-ready
(because the insn queue is full, for example), the main sequencer does not validate any new instruction.
Unluckily, there is a minimum of 1 cycle latency between the validation of the request and the request
that fills the insn queue, so, for one cycle, the sequencer does not see if the just issued instruction
will fill some queues or not. Example: the sequencer gives one instruction to the VSLDU, then it issues
another instruction to the lanes (VFMACC), and both the lanes and the VSLDU becomes non-ready.
Then, the instruction in the lane finishes before the instruction of the VSLDU, but the main sequencer
cannot update the output instruction since the lanes become ready again, and the VSLDU is still not ready.
When the VFMACC finishes, it clears its ID from the running instructions,
so that the lane sequencer can sample the incoming request (again...). And the same VFMACC is issued two times.

Current situation:
Each FU has now a counter in the main sequencer, to keep the count of the instructions in each unit.
The sequencer now can issue instructions if the FUs have space in their queues (check with the counters)
and if the operand requesters are ready.
The instructions are now registered by the main sequencer upon entry
(i.e, the target coutner is updated). In this way, the counter update
does not depend anymore on input signals.
The dispatcher issues a unique token to mark every instruction. Odd and
even instructions have different tokens. The sequencer register an
instruction upon entry when the token is different from the previously
registered one.
Normally, instructions can be issued to the lane sequencer only if
the counters have not reached their maximum capacity.
When an instruction enters the main sequencer, it can happen that the
counter is already at maximum capacity The instruction is
registered anyway taking the counter beyond the maximum capacity.
In this case, the instruction will get a gold ticket, to witness that
it was already registered with the counter, so that the instruction can
pass the checks when the counter returns within its limits, even if
it is at its maximum capacity (the instruction was already counted!)
The main sequencer now does not stall when one FU is full but the
incoming instruction targets another non-full FU.